load user and leagues with new http helpers
