+++ client/app/js/actions/LoadActions.js
@@ -1,32 +1,18 @@
-import {Positions} from '../Constants';
-import buildAsyncAction, {handleAsyncAction, AUTH_REQUIRED} from './buildAsyncAction';
+import buildAsyncAction, {AUTH_REQUIRED} from './buildAsyncAction';
 import {
   LOAD_FANTASY_PLAYERS,
   LOAD_FANTASY_TEAMS,
-  LOAD_FOOTBALL_PLAYERS,
-  LOAD_MY_LEAGUES,
-  LOAD_USER
+  LOAD_FOOTBALL_PLAYERS
 } from './ActionTypes';
 import {
-  selectCurrentUserMeta,
   selectLeagueFantasyPlayersMeta,
   selectLeagueFantasyTeamsMeta,
-  selectLeagueFootballPlayersMeta,
-  selectMyLeaguesMeta
+  selectLeagueFootballPlayersMeta
 } from '../selectors/metaSelectors';
 
-const TEMPTEMP_HARDCODED_LEAGUE_RULES = {
-  max_team_size: 11,
-  team_reqs: {
-    [Positions.QB]: 1,
-    [Positions.RB]: 2,
-    [Positions.WR]: 2,
-    [Positions.TE]: 1,
-    [Positions.FLEX]: 1,
-    [Positions.K]: 1,
-    [Positions['D/ST']]: 1
-  }
-};
+// IN PROGRESS OF MOVING ALL THESE FUNCTIONS TO LoadActions2
+export const loadMyLeagues = require('./LoadActions2').loadMyLeagues;
+export const loadUserAndLeagues = require('./LoadActions2').loadUserAndLeagues;
 
 export function loadFantasyPlayers(fantasyLeagueId) {
   return buildAsyncAction({
@@ -62,37 +48,4 @@ export function loadFootballPlayers(fantasyLeagueId) {
       return selectLeagueFootballPlayersMeta(state, fantasyLeagueId);
     }
   });
-}
-
-export function loadMyLeagues() {
-  return function (dispatch, getState) {
-    const user = selectCurrentUserMeta(getState());
-    if (!user || !user.id) return false;
-
-    return handleAsyncAction(dispatch, getState, {
-      actionType: LOAD_MY_LEAGUES,
-      auth: AUTH_REQUIRED,
-      url: `/api/users/${user.id}/fantasy_leagues/`,
-      parser: parseLeague,
-      metaSelector: selectMyLeaguesMeta
-    });
-  };
-}
-
-export function loadUserAndLeagues() {
-  return buildAsyncAction({
-    actionType: LOAD_USER,
-    auth: AUTH_REQUIRED,
-    url: '/api/current_user',
-    metaSelector: selectCurrentUserMeta,
-    onSuccess: function (dispatch, _getState) {
-      dispatch(loadMyLeagues());
-    }
-  });
-}
-
-function parseLeague(league) {
-  league.draft_start_date = new Date(league.draft_start_date);
-  league.rules = { ...TEMPTEMP_HARDCODED_LEAGUE_RULES };
-  return league;
 }
