+++ client/app/js/http/httpGet.js
@@ -1,6 +1,10 @@
 import request from 'superagent';
 
-export default function httpGet({ url, auth = null }) {
+export default function httpGet({ url, auth = null, token = null }) {
+  if (token) {
+    auth = { u: token, p: '' };
+  }
+
   return new Promise(function (resolve, reject) {
     let req = request.get(url);
     req = (auth ? req.auth(auth.u, auth.p) : req);
